Major character physics update structure debugging
