+++ physics-manager.js
@@ -22,7 +22,6 @@ const localVector5 = new THREE.Vector3();
 const localQuaternion = new THREE.Quaternion();
 const localQuaternion2 = new THREE.Quaternion();
 const localMatrix = new THREE.Matrix4();
-const localObject = new THREE.Object3D();
 
 const zeroVector = new THREE.Vector3(0, 0, 0);
 const upVector = new THREE.Vector3(0, 1, 0);
@@ -231,216 +230,18 @@ const getAvatarHeight = () => {
 };
 physicsManager.getAvatarHeight = getAvatarHeight;
 
-physicsManager.getAvatarCapsule = _getAvatarCapsule;
 physicsManager.physicsEnabled = false;
 physicsManager.setPhysicsEnabled = physicsEnabled => {
   physicsManager.physicsEnabled = physicsEnabled;
 };
-const _applyAvatarPhysics = (camera, avatarOffset, cameraBasedOffset, velocityAvatarDirection, updateRig, timeDiffS) => {
-  if (physicsManager.physicsEnabled) {
-    // apply offset
-    camera.position.add(physicsManager.offset);
-    physicsManager.offset.setScalar(0);
-
-    // capsule physics
-    const localPlayer = metaversefileApi.useLocalPlayer();
-    if (!localPlayer.hasAction('sit')) {
-      applyVelocity(camera.position, localPlayer.characterPhysics.velocity, timeDiffS);
-
-      camera.updateMatrixWorld();
-      camera.matrixWorld.decompose(localVector, localQuaternion, localVector2);
-
-      localVector4.copy(avatarOffset);
-      if (cameraBasedOffset) {
-        localVector4.applyQuaternion(localQuaternion);
-      }
-      localVector.add(localVector4);
-      const collision = _collideCapsule(localVector, localQuaternion2.set(0, 0, 0, 1));
-      
-      // avatar facing direction
-      if (velocityAvatarDirection) {
-        const horizontalVelocity = localVector5.set(
-          localPlayer.characterPhysics.velocity.x,
-          0,
-          localPlayer.characterPhysics.velocity.z
-        );
-        if (horizontalVelocity.lengthSq() > 0.001) {
-          localQuaternion.setFromRotationMatrix(
-            localMatrix.lookAt(
-              zeroVector,
-              horizontalVelocity,
-              upVector
-            )
-          );
-        } else {
-          localPlayer.matrixWorld.decompose(localVector4, localQuaternion, localVector5);
-        }
-      }
-
-      const jumpAction = localPlayer.getAction('jump');
-      const _ensureJumpAction = () => {
-        if (!jumpAction) {
-          const newJumpAction = {
-            type: 'jump',
-            time: 0,
-          };
-          localPlayer.addAction(newJumpAction);
-        } else {
-          jumpAction.set('time', 0);
-        }
-      };
-      const _ensureNoJumpAction = () => {
-        localPlayer.removeAction('jump');
-      };
-      if (collision) {
-        const crouchOffset = getAvatarHeight() * (1 - getPlayerCrouchFactor(localPlayer)) * 0.5;
-        localVector4
-          .fromArray(collision.direction)
-          .add(localVector5.set(0, -crouchOffset, 0));
-        camera.position.add(localVector4)
-        localVector.add(localVector4);
-        
-        if (collision.grounded) {
-          localPlayer.characterPhysics.velocity.y = 0;
-          _ensureNoJumpAction();
-        } else if (!jumpAction) {
-          _ensureJumpAction();
-        }
-      } else if (!jumpAction && localPlayer.characterPhysics.velocity.y < -4) {
-        _ensureJumpAction();
-      }
-    } else {
-      localPlayer.characterPhysics.velocity.y = 0;
-
-      const sitAction = localPlayer.getAction('sit');
-
-      const objInstanceId = sitAction.controllingId;
-      const controlledApp = metaversefileApi.getAppByInstanceId(objInstanceId);
-      const sitPos = sitAction.controllingBone ? sitAction.controllingBone : controlledApp;
-
-      const sitComponent = controlledApp.getComponent('sit');
-      const {sitOffset = [0, 0, 0], damping} = sitComponent;
-
-      physicsManager.setSitOffset(sitOffset);
-
-      applyVelocity(controlledApp.position, localPlayer.characterPhysics.velocity, timeDiffS);
-      if (localPlayer.characterPhysics.velocity.lengthSq() > 0) {
-        controlledApp.quaternion
-          .setFromUnitVectors(
-            localVector4.set(0, 0, -1),
-            localVector5.set(localPlayer.characterPhysics.velocity.x, 0, localPlayer.characterPhysics.velocity.z).normalize()
-          )
-          .premultiply(localQuaternion2.setFromAxisAngle(localVector3.set(0, 1, 0), Math.PI));
-      }
-      controlledApp.updateMatrixWorld();
-
-      localMatrix.copy(sitPos.matrixWorld)
-        .decompose(localVector, localQuaternion, localVector2);
-      
-      localVector.add(_sitOffset);
-      localVector.y += 1;
-      localQuaternion.premultiply(localQuaternion2.setFromAxisAngle(localVector3.set(0, 1, 0), Math.PI));
-      
-      const offset = physicsManager.getAvatarCameraOffset();
-      camera.position.copy(localVector)
-        .sub(
-          localVector3.copy(offset)
-            .applyQuaternion(camera.quaternion)
-        );
-    }
-    localMatrix.compose(localVector, localQuaternion, localVector2);
-
-    // apply to player
-    if (updateRig) {
-      localPlayer.matrix.copy(localMatrix);
-    } else {
-      localPlayer.matrix.identity();
-    }
-    localPlayer.matrix
-      .decompose(localPlayer.position, localPlayer.quaternion, localPlayer.scale);
-    localPlayer.matrixWorld.copy(localPlayer.matrix);
-
-    
-    if (localPlayer.avatar) {
-      if (localPlayer.hasAction('jump')) {
-       localPlayer.avatar.setFloorHeight(-0xFFFFFF);
-      } else {
-        localPlayer.avatar.setFloorHeight(localVector.y - getAvatarHeight());
-      }
-    }
-  }
-};
-const _getAvatarCapsule = v => {
-  const avatarHeight = getAvatarHeight();
-  v.set(0, -avatarHeight * 0.5, 0); // XXX use the proper crouch height
-  v.radius = 0.3/1.6 * avatarHeight;
-  v.halfHeight = Math.max(avatarHeight * 0.5 - v.radius, 0);
-  return v;
-};
-const _collideCapsule = (() => {
-  const localVector = new THREE.Vector3();
-  return (p, q) => {
-    _getAvatarCapsule(localVector);
-    localVector.add(p);
-    return physx.physxWorker.collidePhysics(physx.physics, localVector.radius, localVector.halfHeight, localVector, q, 4);
-  };
-})();
 
 const gravity = new THREE.Vector3(0, -9.8, 0);
 physicsManager.getGravity = () => gravity;
 
-const _copyPQS = (dst, src) => {
-  dst.position.copy(src.position);
-  dst.quaternion.copy(src.quaternion);
-  dst.scale.copy(src.scale);
-};
 const _updatePhysics = timeDiff => {
-  const timeDiffS = timeDiff / 1000;
   const localPlayer = metaversefileApi.useLocalPlayer();
-  const renderer = getRenderer();
-  const session = renderer.xr.getSession();
-
-  const avatarWorldObject = localObject;
-  avatarWorldObject.matrix.copy(camera.matrixWorld)
-    .decompose(avatarWorldObject.position, avatarWorldObject.quaternion, avatarWorldObject.scale);
-  const avatarCameraOffset = session ? zeroVector : cameraManager.getCameraOffset();
-
+  const timeDiffS = timeDiff/1000;
   localPlayer.updatePhysics(timeDiffS);
-
-  if (session) {
-    if (ioManager.currentWalked || localPlayer.hasAction('jump')) {
-      const originalPosition = avatarWorldObject.position.clone();
-
-      _applyAvatarPhysics(avatarWorldObject, avatarCameraOffset, false, false, false, timeDiffS);
-
-      dolly.position.add(
-        avatarWorldObject.position.clone().sub(originalPosition)
-      );
-    } else {
-      localPlayer.characterPhysics.velocity.y = 0;
-    }
-  } else {
-    const selectedTool = cameraManager.getMode();
-    if (selectedTool === 'firstperson') {
-      _applyAvatarPhysics(avatarWorldObject, avatarCameraOffset, true, false, true, timeDiffS);
-      _copyPQS(camera, avatarWorldObject);
-      camera.updateMatrixWorld();
-    } else if (localPlayer.hasAction('aim') && !localPlayer.hasAction('narutoRun')) {
-      _applyAvatarPhysics(avatarWorldObject, avatarCameraOffset, true, false, true, timeDiffS);
-      _copyPQS(camera, avatarWorldObject);
-      camera.updateMatrixWorld();
-    } else if (selectedTool === 'isometric') {
-      _applyAvatarPhysics(avatarWorldObject, avatarCameraOffset, true, true, true, timeDiffS);
-      _copyPQS(camera, avatarWorldObject);
-      camera.updateMatrixWorld();
-    /* } else if (selectedTool === 'birdseye') {
-      _applyAvatarPhysics(avatarWorldObject, avatarCameraOffset, false, true, true, timeDiffS);
-      _copyPQS(camera, avatarWorldObject);
-      camera.updateMatrixWorld(); */
-    } else {
-      throw new Error('invalid camera mode: ' + selectedTool);
-    }
-  }
 };
 physicsManager.update = _updatePhysics;
