+++ util.js
@@ -642,3 +642,9 @@ export function fitCameraToBox(camera, boundingBox, fitOffset = 1) {
 export function applyVelocity(position, velocity, timeDiffS) {
   position.add(localVector.copy(velocity).multiplyScalar(timeDiffS));
 }
+
+export function copyPQS(dst, src) {
+  dst.position.copy(src.position);
+  dst.quaternion.copy(src.quaternion);
+  dst.scale.copy(src.scale);
+}
